# perf: improve the performance of thread_safe_queue