Switch to the usage of a contiguous storage rather than doubly linked lists. The new implementation is still not lock-free. The following benchmarks were completed on AMD Ryzen Threadripper 2920X 12-Core Processor.

Before

```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
thread_safe_queue_push/threads:1         135 ns          135 ns      4590860
thread_safe_queue_push/threads:2         226 ns          378 ns      1852874
thread_safe_queue_push/threads:4         647 ns         1994 ns       386920
thread_safe_queue_push/threads:8        1078 ns         7160 ns       112248
thread_safe_queue_push/threads:12       1142 ns        11215 ns        65556
thread_safe_queue_mpmc/threads:2         243 ns          413 ns      1679348
thread_safe_queue_mpmc/threads:4         623 ns         1972 ns       272808
thread_safe_queue_mpmc/threads:8        1164 ns         7284 ns       138952
thread_safe_queue_mpmc/threads:12       1141 ns        10495 ns        58380
```

After:

```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
thread_safe_queue_push/threads:1        49.7 ns         49.7 ns     14129683
thread_safe_queue_push/threads:2         222 ns          420 ns      2675590
thread_safe_queue_push/threads:4         306 ns         1032 ns       735052
thread_safe_queue_push/threads:8         432 ns         2901 ns       251744
thread_safe_queue_push/threads:12        475 ns         4896 ns       157848
thread_safe_queue_mpmc/threads:2         153 ns          290 ns      2000000
thread_safe_queue_mpmc/threads:4         273 ns          967 ns       711848
thread_safe_queue_mpmc/threads:8         377 ns         2597 ns       260416
thread_safe_queue_mpmc/threads:12        429 ns         4337 ns       120000
```